+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * don't apply formatters for greetings not only for signture (user can use formatting because he control the string)
 * fix max call stack exception when error happen in onEchoCommand
 * Chinese character occupy 2 characters same as in linux terminal (requirement wcwidth and css variables)
+* fix substring and string like '<a' that was breaking command line
+
 
 ## 1.9.0
 ### Features